+++ tests/v2/04_runner/fractal_tasks_core_mock/__init__.py
@@ -1,12 +1,12 @@
+from .converters import create_ome_zarr
+from .converters import create_ome_zarr_multiplex
+from .converters import yokogawa_to_zarr
 from .illumination_correction import illumination_correction
 from .illumination_correction import illumination_correction_B
 from .illumination_correction import init_channel_parallelization
-from .non_parallel_tasks import create_ome_zarr
-from .non_parallel_tasks import create_ome_zarr_multiplex
-from .non_parallel_tasks import new_ome_zarr
+from .mip import maximum_intensity_projection
+from .mip import new_ome_zarr
 from .parallel_tasks import cellpose_segmentation
-from .parallel_tasks import maximum_intensity_projection
-from .parallel_tasks import yokogawa_to_zarr
 from .registration_tasks import apply_registration_to_image
 from .registration_tasks import calculate_registration
 from .registration_tasks import find_registration_consensus
